Updating ShadeBot to 3.0
REWORKED EVERYTHING.
Files and Events are now all seperate files allowing for easier access and edits to commands. A whole new command management allowing for easy to make Alias, Enabled/Disible overall or for Guild only access.

This also fixes some of the stupid little bugs that have been around for a while on the previous version!